vayu: thermal-engine: Count for max value instead of weighted average
Battery is rarely higher than the pmic temp anyway (which is why Xiaomi
didn't use it as a trip point like we do).

So we can just use MATH_MAX (2) instead of the MATH_WEIGHT (0) for the
virtual sensor calculation just in case battery temp goes higher than
the pmic temp.